move version.py to models
