+++ conda/resolve.py
@@ -10,8 +10,8 @@ from .exceptions import NoPackagesFoundError, UnsatisfiableError
 from .logic import Clauses, minimal_unsatisfiable_subset
 from .models.dist import Dist
 from .models.match_spec import MatchSpec
+from .models.version import normalized_version
 from .toposort import toposort
-from .version import normalized_version
 
 log = logging.getLogger(__name__)
 dotlog = logging.getLogger('dotupdate')
